💾 db: replace Upstash cache with IORedis and add Redis config
