+++ apps/server/.env.example
@@ -14,6 +14,8 @@ SUPABASE_SERVICE_ROLE=
 SUPABASE_DATABASE_URL=
 SUPABASE_DATABASE_PASSWORD=
 
+REDIS_URL=
+
 DROPBOX_REFRESH_TOKEN=
 DROPBOX_APP_KEY=
 DROPBOX_APP_SECRET=
