+++ apps/server/src/constants/env.ts
@@ -8,6 +8,7 @@ const env = {
   supabaseUrl: process.env.SUPABASE_URL,
   supabaseKey: process.env.SUPABASE_ANON_KEY,
   databaseUrl: process.env.SUPABASE_DATABASE_URL,
+  redisUrl: process.env.REDIS_URL,
   inngestEventKey: process.env.INNGEST_EVENT_KEY,
   inngestSigningKey: process.env.INNGEST_SIGNING_KEY,
   stardustWebUrl: process.env.STARDUST_WEB_URL,
@@ -33,6 +34,7 @@ const envSchema = z.object({
   supabaseUrl: z.string().url(),
   supabaseKey: z.string(),
   databaseUrl: z.string().url(),
+  redisUrl: z.string().url(),
   inngestEventKey: z.string().optional(),
   inngestSigningKey: z.string().optional(),
   dropboxRefreshToken: z.string(),
