+++ docker-compose.yml
@@ -1,0 +1,11 @@
+services:
+  redis:
+    image: redis:7-alpine
+    container_name: stardust-redis
+    ports:
+      - "6379:6379"
+    volumes:
+      - stardust-redis-data:/data
+
+volumes:
+  stardust-redis-data:
